% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runpoll_3seasons.R
\name{runpoll_3seasons}
\alias{runpoll_3seasons}
\title{Pollinator model function}
\usage{
runpoll_3seasons(
  M_poll0 = numeric(0),
  firstyear,
  firstyearfactor = c(1, 1),
  bees = c("Bombus spp.", "Solitary spp."),
  cell.size = 25,
  paramList,
  nest,
  floral,
  cutoff = 0.99,
  loc_managed
)
}
\arguments{
\item{M_poll0}{the initial population sizes to be used only if we are not in the first year}

\item{firstyear}{a boolean identifying whether we run the model for the first time (in which case we need to initialize the model)}

\item{firstyearfactor}{a factor to control the initial population sizes, defaults to 1.}

\item{bees}{a vector of bees identifiers corresponding to those in the parameter files}

\item{cell.size}{the size of a cell in the landscape matrices, in meters (default to 25)}

\item{paramList}{the list of parameters needed to run the model}

\item{nest}{the rasterstack containing the nesting qualities (dimensions should be consistent with the number of bees in the model)}

\item{floral}{the rasterstack containing the floral values (dimensions should be consistent with the number of bees in the model)}

\item{cutoff, }{the percentile at which the dispersal kernel is cut off}

\item{loc_managed, }{the location at which managed pollinators are placed (e.g. apiaries)}
}
\value{
a list containing :
\describe{
\item{M_poll}{the number of queens at the end of each period and at the beginning of the next season}
\item{N_poll}{the number of foraging bees in each period}
\item{flowvis}{the visitation rates per cell in each period}
\item{floral and nest}{the floral and nesting values (the same as in the inputs, but stored in the results for convenience)}
\item{pollres}{the resources gathered by the bees}
}
}
\description{
Uses spatial data on floral and nesting quality to compute flower visitation and population dynamics for wild pollinators for a given year.
The function can be used for mutliple years by running a loop and using as input the output of the previous model (see example).
}
\examples{

nf<-computeFloralNesting(landuseMap=landuse, edgesMap=stack(grassmargins,flowermargins), unitEdges = "sqm", widthEdges=1,
                         landuseCodes, bees=c("GroundNestingBumblebees", "GroundNestingSolitaryBees"), num_floral=3,
                         florNestInfo=parameters$florNestInfo, codeEdges=c(11,21), cell.size = 10,paramList=parameters)
poll<-runpoll_3seasons(M_poll0 = numeric(0), firstyear=TRUE, firstyearfactor = c(1, 1),
            bees = c("GroundNestingBumblebees", "GroundNestingSolitaryBees"), cell.size = 10, paramList=parameters, nest=nf$nest,
            floral=nf$floral, cutoff = 0.99, loc_managed)
}
\references{
Haeussler J, Sahlin U, Baey C, Smith HG, Clough Y (2017) Predicting pollinator population size and pollination ecosystem service responses
to enhancing floral and nesting resources. Ecology and Evolution, 7: 1898-1908.\url{http://dx.doi.org/10.1002/ece3.2765}
Expansion to 3 seasons first used in:
Gardner E, et al. (2020) Reliably predicting pollinator abundance: Challenges of calibrating
process-based ecological models. Methods in Ecology and Evolution. \url{}
}
